feat(vips): NewBlobFromTarget
