+++ vips/target.go
@@ -4,6 +4,7 @@ package vips
 import "C"
 import (
 	"fmt"
+	"github.com/cshum/imagor"
 	"github.com/cshum/imagor/vips/pointer"
 	"io"
 	"runtime"
@@ -15,7 +16,7 @@ type Target struct {
 	writer io.WriteCloser
 	target *C.VipsTargetCustom
 	ptr    unsafe.Pointer
-	lock   sync.Mutex
+	once   sync.Once
 }
 
 func NewTarget(writer io.WriteCloser) *Target {
@@ -34,13 +35,25 @@ func finalizeTarget(target *Target) {
 }
 
 func (s *Target) Close() {
-	s.lock.Lock()
-	if s.ptr != nil {
-		C.clear_target(&s.target)
-		pointer.Unref(s.ptr)
-		s.ptr = nil
-		_ = s.writer.Close()
-		log("vips", LogLevelDebug, fmt.Sprintf("closing target %p", s))
-	}
-	s.lock.Unlock()
+	s.once.Do(func() {
+		if s.ptr != nil {
+			C.clear_target(&s.target)
+			pointer.Unref(s.ptr)
+			s.ptr = nil
+			_ = s.writer.Close()
+			log("vips", LogLevelDebug, fmt.Sprintf("closing target %p", s))
+		}
+	})
+}
+
+func NewBlobFromTarget(handler func(*Target) error) *imagor.Blob {
+	pr, pw := io.Pipe()
+	target := NewTarget(pw)
+	go func() {
+		defer target.Close()
+		if err := handler(target); err != nil {
+			_ = pr.CloseWithError(err)
+		}
+	}()
+	return imagor.NewBlobFromReader(pr)
 }
